update advertiseFilter interface

+++ src/main/java/com/st/BlueNRG/fwUpgrade/BlueNRGAdvertiseFilter.java
@@ -89,7 +89,7 @@ public class BlueNRGAdvertiseFilter implements AdvertiseFilter {
         }
 
         @Override
-        public int getFeatureMap() {
+        public long getFeatureMap() {
             return 0;
         }
